==> notebooks/0_initial_cleaning/2_dataset_cleaning.ipynb <==
import pandas as pd
import numpy as np
import os
import pyspark
from pyspark.sql import SparkSession
from pyspark.sql.types import StructType, StructField, StringType

spark = SparkSession.builder \
    .appName('airlines_delay')\
    .config('spark.jars', '../../jars/snowflake-jdbc-3.13.6.jar, ../../jars/spark-snowflake_2.12-2.9.0-spark_3.1.jar') \
    .getOrCreate()

cols = ['Year',
    'Quarter',
    'Month',
    'DayofMonth',
    'DayOfWeek',
    'FlightDate',
    'Reporting_Airline',
    'DOT_ID_Reporting_Airline',
    'IATA_CODE_Reporting_Airline',
    'Tail_Number',
    'Flight_Number_Reporting_Airline',
    'OriginAirportID',
    'Origin',
    'OriginCityName',
    'OriginState',
    'DestAirportID',
    'Dest',
    'DestCityName',
    'DestState',
    'CRSDepTime',
    'DepTime',
    'DepDelay',
    'DepDelayMinutes',
    'CRSArrTime',
    'ArrTime',
    'ArrDelay',
    'ArrDelayMinutes',
    'Cancelled',
    'CancellationCode',
    'Diverted',
    'CRSElapsedTime',
    'ActualElapsedTime',
    'AirTime',
    'Flights',
    'Distance',
    'CarrierDelay',
    'WeatherDelay',
    'NASDelay',
    'SecurityDelay',
    'LateAircraftDelay'
]

file_path = "../../datasets/t_ontime_small.csv"

delay = spark.read.csv(file_path,header=False)
delay = delay.toDF(*cols)

delay

delay.limit(20).toPandas()

# Same Reporting Airline and IATA Code - One can be removed
delay_same_code = delay.filter(delay["Reporting_Airline"] != delay["IATA_CODE_Reporting_Airline"])
delay_same_code.show()

# Total 21 Airlines
delay.select('Reporting_Airline').distinct().collect()

# Analyzing Airport Related Columns
airport_cols = ["OriginAirportID", "Origin", "OriginCityName", "OriginState", "DestAirportID", "Dest", "DestCityName", "DestState"]
airport_info_selected = delay.select(*airport_cols).show()

# Duplicate Data because of Airports specific Column. The distinct airport can be moved to another DF.
# Apart from OriginAirportID and DestAirportID, other columns can be removed.
origin_airport_cols = ["OriginAirportID", "Origin", "OriginCityName", "OriginState"]
origin_airport_info_selected = delay.select(*origin_airport_cols)

dest_airport_cols = ["DestAirportID", "Dest", "DestCityName", "DestState"]
dest_airport_info_selected = delay.select(*dest_airport_cols)

emptyRDD = spark.sparkContext.emptyRDD()
schema = StructType([
  StructField('AirportID', StringType(), True),
  StructField('CityCode', StringType(), True),
  StructField('CityName', StringType(), True),
  StructField('StateCode', StringType(), True)
  ])
airport_df = spark.createDataFrame(emptyRDD,schema)
airport_df=airport_df.union(origin_airport_info_selected)
airport_df=airport_df.union(dest_airport_info_selected)

airport_df = airport_df.distinct()
print("Distinct Airports: "+str(airport_df.count()))
airport_df.show(truncate=False)
airport_df.repartition(1).write.csv("../../datasets/airports.csv")

# Check Data with Code US, FL and VX - Code not found on the official IATA website
delay.select('IATA_CODE_Reporting_Airline').distinct().collect()

# Casting the FlightDate column as Date. It would help with future computations
delay = delay.withColumn("FlightDate", delay["FlightDate"]
                                  .cast('date'))

delay
