==> notebooks/0_initial_cleaning/5_airports_to_sqlite.ipynb <==
## Imports

import pandas as pd
import numpy as np
import sqlite3
import sqlalchemy
from sqlalchemy import create_engine

file_path = "../../datasets/airports.csv"

df = pd.read_csv(file_path, names=['AirportID', 'Airport', 'City', 'State'])

df

def split_city(val):
    return val.split(",")[0]

df['City'] = df.City.apply(split_city)

df

df_airports = pd.read_csv("https://raw.githubusercontent.com/mborsetti/airportsdata/main/airportsdata/airports.csv")

df_airports

airports_final = df.set_index('Airport').join(df_airports.set_index('iata'))[['AirportID', 'City', 'State', 'elevation', 'lat', 'lon']].reset_index().rename({"index":"Airport", "elevation": "Elevation", "lat":"Lat", "lon":"Lon"}, axis=1)

airports_final

airports_final.to_csv('../../datasets/airports_final.csv', index=False)

db_path = "../../dbs/delays.db"

conn = sqlite3.connect(db_path)
conn.execute("DROP TABLE IF EXISTS airport;")

create_query = ''' \
            CREATE TABLE airport
            ( 
                Airport                         VARCHAR, 
                AirportID                       INT,
                City                            VARCHAR,
                State                           VARCHAR,
                Elevation                       REAL,
                Lat                             REAL,
                Lon                             REAL
            );
    '''
conn.execute(create_query)
conn.close()
engine = sqlalchemy.create_engine("sqlite:///"+db_path, echo=False)

engine = sqlalchemy.create_engine("sqlite:///"+db_path, echo=False)
airports_final.to_sql('airport', con=engine, if_exists='append', index=False)
