==> notebooks/0_initial_cleaning/6_air_carriers.ipynb <==
import pandas as pd
import numpy as np
import sqlite3
import sqlalchemy
from sqlalchemy import create_engine

file_path = "../../datasets/air_carriers.csv"

df = pd.read_csv(file_path, names=['IATACode', 'ICAOCode', 'AirCarrierName'], header=0)

air_carriers = df.set_index('IATACode')

air_carriers

db_path = "../../dbs/delays.db"

conn = sqlite3.connect(db_path)
conn.execute("DROP TABLE IF EXISTS air_carriers;")

create_query = ''' \
            CREATE TABLE air_carriers
            ( 
                IATACode                       VARCHAR, 
                ICAOCode                       VARCHAR,
                AirCarrierName                 VARCHAR
            );
    '''
conn.execute(create_query)
conn.close()
engine = sqlalchemy.create_engine("sqlite:///"+db_path, echo=False)

engine = sqlalchemy.create_engine("sqlite:///"+db_path, echo=False)
air_carriers.to_sql('air_carriers', con=engine, if_exists='append', index=False)
